refactor personending
